backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=2.0.0
redis>=5.0.0
playwright>=1.42.0
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0
pyotp>=2.9.0
jinja2>=3.1.2
aiosmtplib>=2.0.2
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.2
pytest>=7.4.0
httpx>=0.25.0
requests>=2.31.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
elasticsearch
